Boot_disk allows a mirage kernel to be uploaded as a bootable VDI
It synthesises a disk image containing an MBR and a single partition
containing a FAT filesystem. The kernel goes into the filesystem
together with grub config. The disk image is then uploaded direct
to a XenServer using the import_raw_vdi interface. Pygrub is then
able to boot PV VMs from this disk.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>